+++ tests/test_benchmark_script.py
@@ -41,6 +41,7 @@ class TestBenchmarkScript:
             "shortest_path_grid_120",
             "nearest_node_grid_2000",
             "export_geojson_grid_120_compact",
+            "export_bundle_json_grid_120_compact",
             "export_bundle_end_to_end",
         }